Research pivot: config comparison, E2E latency, YOLO CPU baseline
Advisor feedback (March 23): shift from isolated benchmarks to
system-level configuration comparison with E2E latency as headline metric.

YOLO + ViSP (Ahmed):
- visp_servo_node.cpp: add detection-header-stamp → cmd_vel E2E latency
  measurement; logs 'e2e=XXms' alongside existing callback latency
- yolo_cpu_producer.py: new CPU inference producer (Ultralytics PyTorch);
  drop-in replacement for Hailo NPU path, same SHM interface
- benchmarks/e2e_latency_probe.py: ROS2 probe node for det→cmd_vel latency
- benchmarks/contention_bench.sh: perf stat profiling across 4 conditions
- benchmarks/yolo_cpu_bench.sh: benchmark all YOLO26 sizes on CPU

SLAM (Amar):
- ov2slam_ros: precision_config.hpp, frame.hpp/cpp float32 bearing vectors
- FE speedup configs: fe_klt2, fe_maxdist65, fe_combined yamls
- aggregate_slam_results.py, plot_paper_figures.py, plot_ipc_cdf.py,
  core_pinning_bench.sh, ipc_latency_bench.py, plot_resolution.py,
  yolo_fp_breakdown.sh, yolo_nms_comparison.sh, yolo_resolution_bench.sh

Cleanup:
- Remove gstshark calibration trace (camera_calib/)
- Add benchmarks/results/ and .claude/ to .gitignore

+++ src/ov2slam_ros/src/mapper.cpp
@@ -295,13 +295,13 @@ void Mapper::triangulateTemporal(Frame &frame)
         }
 
         // Check rotation-compensated parallax
-        cv::Point2f rotpx = frame.projCamToImage(Rcicj * vkps.at(i).bv_);
+        cv::Point2f rotpx = frame.projCamToImage(Rcicj * vkps.at(i).bv_.template cast<double>());
         double parallax = cv::norm(kfkp.unpx_ - rotpx);
 
         candidates++;
 
         // Compute 3D pos and check if its good or not
-        left_pt = computeTriangulation(Tcicj, kfkp.bv_, vkps.at(i).bv_);
+        left_pt = computeTriangulation(Tcicj, kfkp.bv_.template cast<double>(), vkps.at(i).bv_.template cast<double>());
 
         // Project into right cam (new KF)
         right_pt = Tcjci * left_pt;
@@ -383,8 +383,8 @@ void Mapper::triangulateStereo(Frame &frame)
         if( !vkps.at(i).is3d_ && vkps.at(i).is_stereo_ )
         {
             vstereoidx.push_back(i);
-            vleftbvs.push_back(vkps.at(i).bv_);
-            vrightbvs.push_back(vkps.at(i).rbv_);
+            vleftbvs.push_back(vkps.at(i).bv_.template cast<double>());
+            vrightbvs.push_back(vkps.at(i).rbv_.template cast<double>());
         }
     }
